Refactor initialization API to conform to shared design with iOS SDK.
